fix remaining clang-tidy findings in the C sources
Fix the two findings left in issue #353:

h_malloc.c:432 clang-analyzer-security.ArrayBound: false positive.
slots is limited to MAX_SLAB_SLOT_COUNT (256) via size_class_slots, so
(slots - 1) / U64_WIDTH <= 3 and i always stays within the 4 bitmap
words; random_index is bounded by slots as well. The analyzer can't see
either bound across translation units (slots is loaded from a u16 field
and get_random_u16_uniform is defined in random.c), so suppress the
report with a targeted NOLINT instead of disabling the check.

util.h:51 bugprone-narrowing-conversions: make the conversion explicit
with a cast. int64_t rather than long long since casting to the higher
rank long long trips bugprone-misplaced-widening-cast. The same bits
reach __builtin_ffsll as before and the built library is unchanged.

+++ util.h
@@ -48,7 +48,7 @@ typedef unsigned __int128 u128;
 #define U64_WIDTH 64
 
 static inline int ffz64(u64 x) {
-    return __builtin_ffsll(~x);
+    return __builtin_ffsll((int64_t) ~x);
 }
 
 // parameter must not be 0
